+++ src/tests/test_skip_connection_mutation.py
@@ -5,7 +5,7 @@ import torch.nn.functional as F
 import torch.fx
 from torch.fx.passes.shape_prop import ShapeProp
 from src.utils import find_required_shapes, add_skip_connection
-from src.modelbank import simple_linear
+from src.tests.modelbank import simple_linear
 from src.core import get_graph, add_node
 import random
